[#18] Creates UpfBpfConfig file useful to use with find_package
Usage:
$CMAKE $cmake_args .. -DUpfBpf_DIR="${THIS_SCRIPT_PATH}/../ext/upf-bpf/package/lib/cmake/UpfBpf/

Adds in CMakeList.txt

```
find_package(UpfBpf CONFIG REQUIRED)
if(NOT TARGET spdlog)
    # Stand-alone build
    find_package(spdlog REQUIRED)
    message("SPDLOG package found.")
endif()
message("UpfBpf_DIR: ${UpfBpf_DIR}")
```

Notice that we depend on spdlog.
TODO - Encapsulate with UpfBpf package.

There was a problem during the trial with openair-cn-cups which also uses
the spdlog.

- Firstly, because it was used different spdlog version
- Secondly, because it was used different spdlog header

It must be check deeper.